+++ gl/include/BoundingBoxRenderer.h
@@ -14,7 +14,7 @@ public:
   virtual void initGL();
   virtual void updateSpins(const std::vector<glm::vec3>& positions,
                            const std::vector<glm::vec3>& directions);
-  virtual void draw(double aspectRatio) const;
+  virtual void draw(float aspectRatio) const;
 
 protected:
   virtual void optionsHaveChanged(const std::vector<int>& changedOptions);
