+++ gl/include/Camera.h
@@ -11,22 +11,22 @@ public:
   Camera(const glm::vec3& camera_position,
          const glm::vec3& center_position,
          const glm::vec3& up_vector,
-         const double& aspect_ratio);
+         const float& aspect_ratio);
 
-  void setAspectRatio(double aspect_ratio);
+  void setAspectRatio(float aspect_ratio);
   
   void lookAt(glm::vec3 eye, glm::vec3 center, glm::vec3 up);
   
   const glm::vec3& cameraPosition() const;
   const glm::vec3& centerPosition() const;
   const glm::vec3& upVector() const;
-  const double& aspectRatio() const;
+  const float& aspectRatio() const;
   
 private:
   glm::vec3 _camera_position;
   glm::vec3 _center_position;
   glm::vec3 _up_vector;
-  double _aspect_ratio;
+  float _aspect_ratio;
 };
 
 #endif // CAMERA_H
