+++ gl/src/Camera.cpp
@@ -9,7 +9,7 @@ Camera::Camera() : Camera(glm::vec3(0, 0, 1), glm::vec3(0, 0, 0), glm::vec3(0, 1
 Camera::Camera(const glm::vec3& camera_position,
                const glm::vec3& center_position,
                const glm::vec3& up_vector,
-               const double& aspect_ratio) :
+               const float& aspect_ratio) :
 _camera_position(camera_position),
 _center_position(center_position),
 _up_vector(up_vector),
@@ -22,7 +22,7 @@ void Camera::lookAt(glm::vec3 eye, glm::vec3 center, glm::vec3 up) {
   _up_vector = up;
 }
 
-void Camera::setAspectRatio(double aspect_ratio) {
+void Camera::setAspectRatio(float aspect_ratio) {
   _aspect_ratio = aspect_ratio;
 }
 
@@ -38,6 +38,6 @@ const glm::vec3& Camera::upVector() const {
   return _up_vector;
 }
 
-const double& Camera::aspectRatio() const {
+const float& Camera::aspectRatio() const {
   return _aspect_ratio;
 }
